feat: Implement modal management for user account actions
- Added `ModalsRoot` component to manage modals for creating, updating, and viewing user details.
- Created `CreateUserModal`, `DetailsUserModal`, and `UpdateUserModal` components for respective user actions.
- Introduced `DiscloresuresProvider` to manage disclosure states for user modals.
- Implemented hooks for managing modal disclosures: `useCreateUserDisclosure`, `useUpdateUserDisclosure`, and `useViewUserByIdDisclosure`.
- Integrated user management table with search functionality and action icons for user details, editing, and deletion.
