+++ src/components/modules/Modal/ManageAccount/CreateUser/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const CreateUserModal = () => {
+  return (
+    <div>Create User</div>
+  )
+}
+
+export default CreateUserModal
