+++ src/components/modules/Modal/ManageAccount/DetailsUser/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const DetailsUserModal = () => {
+  return (
+    <div>Details User</div>
+  )
+}
+
+export default DetailsUserModal
