+++ src/components/modules/Modal/ManageAccount/UpdateUser/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const UpdateUserModal = () => {
+  return (
+    <div>Update User</div>
+  )
+}
+
+export default UpdateUserModal
